Validate parsers against live FAA data; fix silent drops
Validate every record parser's field comments against the live FAA
distribution, enforce strict swift-format and SwiftLint, and fix the
silent record/field drops that validation surfaced.

+++ Sources/SwiftNASR/Models/Records/Fix.swift
@@ -155,7 +155,7 @@ public struct Fix: ParentRecord {
     public let navaidType: Navaid.FacilityType
 
     /// The radial or bearing from the navaid (degrees).
-    public let radialDeg: UInt?
+    public let radialDeg: Float?
 
     /// The DME distance from the navaid (nautical miles).
     public let distanceNM: Float?
